remove active_fd[] in select
old version:
declare an active_fd, and block_fd in select
once a new link come in, add it to active_fd.
and if the active_fd is select, add it to block_fd to prevent other
thread get it.
In the end of send data, remove self from block_fd.

new version:
server act to login after get a new link from socket.
Though this might prone to error, but I think it's OK.
select check all server->client, and add all client into watching
list. If there is a link come in, status of that client will changes
to PROCESS from LINK. The status in PROCESS will not add to
select list.